(main)のlayout.tsxに初期データの取得とStoreInitializerを移動
